# Introduce StageManager for managing tasks stage by stage