# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #1936, #1704.

 # Rationale for this change
<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->
Previously the task fetching time complexity is n^2. By introducing stage-based scheduling, it can be reduced to n. And it will also be easy for task error recovering and stage priority-based scheduling in the future.

What's more, by this PR, the performance of Pull-Staged task scheduling is also improved a lot due to there's no need to check all of the tasks status in the whole system.

# What changes are included in this PR?
<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->
- Introduced *Stage* and *StageManager*
- Removed *SchedulerStateWatcher*, due to we have introduced event-based processing mechanism for the state machine changes.

Details see #1936.

# Are there any user-facing changes?
<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->
It's blocked by #1934.