@yahoNanJing 2022-03-14
Hi @houqp, @thinkharderdev, @yjshen, I've just fixed the bug for the master branch's push-based task scheduling. Now we can have the whole view of benchmark results.
- For the push-based task scheduling, with this PR, the performance is improved around 2x.
- For the pull-based task scheduling, with this PR, the performance is improved around 2x at the beginning and will not be linearly downgraded.