@yahoNanJing 2022-03-16
> Could we use a cfg to achieve the same result. So something like:

Thanks @thinkharderdev for your suggestion. I used cfg at the beginning. However, I don't know whether this way is acceptable or not. If we all prefer this way, I'll change it back to cfg way.