@yahoNanJing 2022-03-16
> I think the ExecutionContext should not a member of QueryStageScheduler, since we will support session context soon.
> 
> ```
> pub(crate) struct QueryStageScheduler<
>     T: 'static + AsLogicalPlan,
>     U: 'static + AsExecutionPlan,
> > {
>     ctx: Arc<RwLock<SessionContext>>,
>     state: Arc<SchedulerState<T, U>>,
>     event_sender: Option<EventSender<SchedulerServerEvent>>,
> }
> ```

Thanks @mingmwang. It makes sense not to include context to the QueryStateScheduler. Will refactor it.